theoretically working staging / production webpack configs
