+++ .gitignore
@@ -11,6 +11,7 @@
 !/tmp/.keep
 
 # webpacker
+/public/assets
 /public/packs
 /public/packs-test
 /node_modules
