+++ config/webpack/shared.js
@@ -3,33 +3,18 @@ require('./mapEnv')
 const { environment } = require('@rails/webpacker')
 
 const path = require('path')
-const webpack = require('webpack')
 const merge = require('webpack-merge')
 const constants = require('./constants')
 
-function config (preprocess) {
-  if (preprocess) {
-    preprocess(environment)
+// blow away the built-in manifest plugin so that it can be customized
+environment.plugins.delete('Manifest')
+
+const base = merge(environment.toWebpackConfig(), {
+  resolve: {
+    alias: {
+      shared: path.resolve(constants.client, './src/shared')
+    }
   }
+})
 
-  return merge(environment.toWebpackConfig(), {
-    resolve: {
-      alias: {
-        shared: path.resolve(constants.client, './src/shared')
-      }
-    }
-  })
-}
-
-function environmentPlugin (value) {
-  return new webpack.DefinePlugin({
-    'process.env': {
-      'ENVIRONMENT': JSON.stringify(value)
-    }
-  })
-}
-
-module.exports = {
-  config,
-  environmentPlugin
-}
+module.exports = base
